Colog.Core.Class

Description

Provides type class for values that has access to LogAction.

Documentation

class HasLog env msg m where

This types class contains simple pair of getter-setter and related functions. It also provides the useful lens logActionL with the default implementation using type class methods. The default one could be easily overritten under your instances.

Every instance of the this typeclass should satisfy the following laws:

Set-Get: getLogAction (setLogAction l env) ≡ l

Get-Set: setLogAction (getLogAction env) env ≡ env

Set-Set: setLogAction l2 (setLogAction l1 env) ≡ setLogAction l2 env

Set-Over: overLogAction f env ≡ setLogAction (f $ getLogAction env) env

Minimal complete definition

logActionL | getLogAction, (setLogAction | overLogAction)

Methods

getLogAction :: env -> LogAction m msg

Extracts LogAction from the environment.

setLogAction :: LogAction m msg -> env -> env

Sets LogAction to the given one inside the environment.

overLogAction :: (LogAction m msg -> LogAction m msg) -> env -> env

Applies function to the LogAction inside the environment.

logActionL :: Lens' env (LogAction m msg)

Lens for LogAction inside the environment.

Lens

To keep co-log-core a lightweight library it was decided to introduce local Lens' type alias as it doesn't harm.

type Lens' s a = forall f. Functor f => (a -> f a) -> s -> f s

The monomorphic lenses which don't change the type of the container (or of the value inside).