Colog.Core.Severity

Description

This module introduces Severity data type for expressing how severe the message is. Also, it contains useful functions and patterns for work with Severity.

Severity	Pattern	Meaning	Example

Debug	D	Information useful for debug purposes	Internal function call logs

Info	I	Normal operational information	Finish file uploading

Warning	W	General warnings, non-critical failures	Image load error

Error	E	General errors/severe errors	Could not connect to the DB

Documentation

data Severity

Severity for the log messages.

Constructors

Debug
Information useful for debug purposes.

E.g. output of the function that is important for the internal development, not for users. Like, the result of SQL query.

Info
Normal operational information.

E.g. describing general steps: starting application, finished downloading.

Warning
General warnings, non-critical failures.

E.g. couldn't download icon from some service to display.

Error
General errors/severe errors.

E.g. exceptional situations: couldn't syncronize accounts.

Patterns

Instead of using full names of the constructors you can instead use one-letter patterns. To do so you can import and use the pattern:

import Colog (pattern D)

example :: WithLog env Message m => m ()
example = log D "I'm using severity pattern"

Moreover, you could use patterns when pattern-matching on severity

errorToStderr :: Severity -> IO ()
errorToStderr E = hputStrLn stderr "Error severity"
errorToStderr _ = putStrLn "Something else"

pattern D :: Severity

pattern I :: Severity

pattern W :: Severity

pattern E :: Severity

filterBySeverity :: Applicative m => Severity -> (a -> Severity) -> LogAction m a -> LogAction m a

Filters messages by the given Severity.

data WithSeverity msg

A message tagged with a Severity.

It is common to want to log various types of messages tagged with a severity. WithSeverity provides a standard way to do so while allowing the messages to be processed independently of the severity.

It is easy to cmap over a 'LogAction m (WithSeverity a)', or to filter based on the severity.

logSomething :: LogAction m (WithSeverity String) -> m ()
logSomething logger = logger <& "hello" `WithSeverity` Info

cmap' :: (b -> a) -> LogAction m (WithSeverity a) -> LogAction m (WithSeverity b)
cmap' f action = cmap (fmap f) action

filterBySeverity' :: (Applicative m) => Severity -> LogAction m (WithSeverity a) -> LogAction m (WithSeverity a)
filterBySeverity' threshold action = filterBySeverity threshold getSeverity action

Since: 0.3.1.0

Constructors

WithSeverity

Fields

getMsg :: msg

getSeverity :: Severity

mapSeverity :: (Severity -> Severity) -> WithSeverity msg -> WithSeverity msg

Map the given function over the severity of a WithSeverity.

This can be useful to operate generically over the severity, for example:

suppressErrors :: LogAction m (WithSeverity msg) -> LogAction m (WithSeverity msg)
suppressErrors = cmap (mapSeverity (s -> if s == Error then Warning else s))

Since: 0.3.1.0